smart-tactic-test/chat_service/requirements.txt
flask==2.3.3
flask-cors==4.0.0
flask-limiter==3.5.0
pydantic==2.9.2
google-generativeai==0.8.3
tinydb==4.8.0
python-dotenv==1.0.0
